requirements.txt
Django
djangorestframework
mysqlclient
django-mongodb-backend
python-dotenv
djangorestframework-simplejwt
pymongo
django-filter
gunicorn
whitenoise
drf-spectacular
